+++ kotsadm/web/src/components/snapshots/SnapshotStorageDestination.jsx
@@ -7,6 +7,8 @@ import find from "lodash/find";
 
 import "../../scss/components/shared/SnapshotForm.scss";
 
+import SnapshotInstallationBox from "./SnapshotInstallationBox";
+
 const DESTINATIONS = [
   {
     value: "aws",
@@ -589,45 +591,9 @@ class SnapshotStorageDestination extends Component {
             <span className="icon clickable backArrow-icon u-marginRight--10" style={{ verticalAlign: "0" }} />
             Back
           </div>
-        : null}
+          : null}
         <p className="u-fontSize--largest u-marginBottom--20 u-fontWeight--bold u-color--tundora u-marginTop--12">Configure storage destination</p>
         <p className="u-fontSize--normal u-color--dustyGray u-lineHeight--normal u-fontWeight--medium">To begin with snapshots you need to configure where you want them to be stored. Snapshots can be stored on Amazon S3, Google Cloud Storage, Azure Blob Storage, and other S3 compatible storage providers.</p>
-        <div className={`${snapshotSettings?.isVeleroRunning ? "u-display--none" : "flex u-marginTop--12"}`}>
-          <div className="flex u-marginRight--20">
-            <span className="icon redWarningIcon" />
-          </div>
-          <div className="flex flex-column">
-            <p className="u-color--chestnut u-fontSize--larger u-fontWeight--bold"> Velero is not running </p>
-            <p className="u-fontSize--small u-color--dustyGray u-lineHeight--normal u-fontWeight--medium u-marginTop--10">
-              Velero has been detected, but it's not running successfully. Snapshots will not work until Velero is running reliably.
-        <a href="https://kots.io/kotsadm/snapshots/troubleshooting/" target="_blank" rel="noopener noreferrer" className="replicated-link u-marginLeft--5">Get help</a>
-            </p>
-          </div>
-        </div>
-        <div className={`${snapshotSettings?.veleroVersion !== "" && snapshotSettings?.resticVersion === "" ? "flex u-marginTop--12" : "u-display--none"}`}>
-          <div className="flex u-marginRight--20">
-            <span className="icon redWarningIcon" />
-          </div>
-          <div className="flex flex-column">
-            <p className="u-color--chestnut u-fontSize--larger u-fontWeight--bold"> Restic integration not found </p>
-            <p className="u-fontSize--small u-color--dustyGray u-lineHeight--normal u-fontWeight--medium u-marginTop--10">
-              The Admin Console requires the Velero restic integration to use Snapshots, but it was not found. Please install the Velero restic integration to continue.
-        <a href="https://kots.io/kotsadm/snapshots/" target="_blank" rel="noopener noreferrer" className="replicated-link u-marginLeft--5">Get help</a>
-            </p>
-          </div>
-        </div>
-        <div className={`${snapshotSettings?.veleroVersion !== "" && snapshotSettings?.resticVersion !== "" && !snapshotSettings?.isResticRunning ? "flex u-marginTop--12" : "u-display--none"}`}>
-          <div className="flex u-marginRight--20">
-            <span className="icon redWarningIcon" />
-          </div>
-          <div className="flex flex-column">
-            <p className="u-color--chestnut u-fontSize--larger u-fontWeight--bold"> Restic is not working </p>
-            <p className="u-fontSize--small u-color--dustyGray u-lineHeight--normal u-fontWeight--medium u-marginTop--10">
-              Velero and the restic integration have been detected, but restic is not running successfully. Snapshots will not work until Restic is running reliably.
-        <a href="https://kots.io/kotsadm/snapshots/restic-troubleshooting/" target="_blank" rel="noopener noreferrer" className="replicated-link u-marginLeft--5">Get help</a>
-            </p>
-          </div>
-        </div>
         <div className="flex u-marginTop--20">
           <form className="flex flex-column snapshot-form-wrapper u-marginRight--50">
             {updateErrorMsg &&
@@ -668,17 +634,12 @@ class SnapshotStorageDestination extends Component {
             }
           </form>
 
-          <div className="CheckVelero--wrapper flex1 flex-column justifyContent--center">
-            <p className="u-color--tundora u-fontSize--large u-fontWeight--bold">Check Velero installation</p>
-            {!this.props.hideCheckVeleroButton ?
-              <div className="u-marginTop--12">
-                <button className="btn secondary blue" onClick={() => this.props.fetchSnapshotSettings(true)}>Check for Velero</button>
-              </div>
-              : this.props.renderNotVeleroMessage()
-            }
-            {snapshotSettings.veleroVersion !== "" ?
-              <span className="flex alignItems--center u-marginTop--10 u-fontSize--small u-fontWeight--medium u-color--tuna"><span className="icon checkmark-icon u-marginRight--5" />Velero is installed on your cluster</span> : null}
-          </div>
+          <SnapshotInstallationBox
+            fetchSnapshotSettings={this.props.fetchSnapshotSettings}
+            renderNotVeleroMessage={this.props.renderNotVeleroMessage}
+            snapshotSettings={snapshotSettings}
+            hideCheckVeleroButton={this.props.hideCheckVeleroButton}
+          />
         </div>
       </div>
     );
